requirements.txt
deepgram-sdk==3.7.4
python-dotenv
groq
ffmpeg-python
